TLE using data from http-request node
In this example, an http-request node is used to get the TLE data for a satellite. The string that comes back is split and moved into msg.satid, msg.tel1, msg.tel2 and msg.payload is set to a timestamp. The messages are tehn sent to a join to create one msg that goes to the TLE node to get the coordinates of the sattelite.