Example of an http-request node feeding the TLE node
This example uses an 'http-request' to get the satellite id, tle1 and tle2 data. The results come back as a '\n' seperated string.  The msg is fed into a 'split' node and then to a 'switch' node. The 'switch' node sends each of the four parts - coming from the 'join' - to seperate 'change' nodes to move the payload to the correct msg variable . For example the first msg part is the satellite ID so it is moved to msg.satid. 

All four msgs feed a 'join' node and then to another 'change' node that sets msg.payload to a timestamp. Finally the msg is feed into the 'TLE' node to get the results.